move WeightsTorch class to utils to avoid dependence of profiling tools on torch
